service/requirements.txt
dbus-python>=1.3.2
PyGObject>=3.42.0
pyudev>=0.24.0
psutil>=5.9.0
toml>=0.10.2

